backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
celery==5.3.1
redis==4.5.1

numpy==1.24.4
opencv-python-headless==4.8.1.78
Pillow==10.0.0

pytesseract==0.3.10
torch==2.2.0
transformers==4.35.2

python-multipart==0.0.6
rapidfuzz==3.8.1
python-Levenshtein==0.21.1
reportlab==3.6.12

spacy==3.7.2
